requirements.txt
Flask
Flask-SQLAlchemy
pymysql
jsonschema